Organize perturbation schemes (#6)
* Move perturbation schemes into their own module and simplify code

* Improve Simulated Annealing code and tests
