+++ python_tsp/heuristics/simulated_annealing.py
@@ -1,8 +1,9 @@
-from typing import Callable, Dict, Generator, List, Optional, Tuple
+from typing import List, Optional, Tuple
 
 import numpy as np
 
-from python_tsp.heuristics import local_search
+from python_tsp.heuristics.perturbation_schemes import neighborhood_gen
+from python_tsp.heuristics.local_search import setup
 from python_tsp.utils import compute_permutation_distance
 
 
@@ -57,8 +58,8 @@ def solve_tsp_simulated_annealing(
     339-355.
     """
 
-    x, fx = local_search._setup(distance_matrix, x0)
-    temp = initial_temperature(distance_matrix, x, fx, perturbation_scheme)
+    x, fx = setup(distance_matrix, x0)
+    temp = _initial_temperature(distance_matrix, x, fx, perturbation_scheme)
 
     n = len(x)
     k_inner_min = 12 * n  # min inner iterations
@@ -71,7 +72,7 @@ def solve_tsp_simulated_annealing(
             xn = _perturbation(x, perturbation_scheme)
             fn = compute_permutation_distance(distance_matrix, xn)
 
-            if acceptance_rule(fx, fn, temp):
+            if _acceptance_rule(fx, fn, temp):
                 x, fx = xn, fn
                 k_accepted += 1
                 k_noimprovements = 0
@@ -95,7 +96,7 @@ def solve_tsp_simulated_annealing(
     return x, fx
 
 
-def initial_temperature(
+def _initial_temperature(
     distance_matrix: np.ndarray,
     x: List[int],
     fx: float,
@@ -140,18 +141,10 @@ def _perturbation(x: List[int], perturbation_scheme: str = "ps3"):
     module, and pick the first solution. Since the neighborhood is randomized,
     it is the same as creating a random perturbation.
     """
-    neighborhood_gen: Dict[
-        str, Callable[[List[int]], Generator[List[int], List[int], None]]
-    ] = {
-        "ps1": local_search.ps1_gen,
-        "ps2": local_search.ps2_gen,
-        "ps3": local_search.ps3_gen,
-    }
-
     return next(neighborhood_gen[perturbation_scheme](x))
 
 
-def acceptance_rule(fx: float, fn: float, temp: float) -> bool:
+def _acceptance_rule(fx: float, fn: float, temp: float) -> bool:
     """Metropolis acceptance rule"""
 
     dfx = fn - fx
