+++ tests/test_local_search.py
@@ -20,37 +20,6 @@ class TestImportModule:
         )
 
 
-class TestPerturbationSchemes:
-    x = [0, 1, 2, 3, 4]
-
-    def test_ps1_returns_correct_num_neighbors(self):
-        """PS1 has n - 1 swaps.
-        But since we fix the first element as origin, it leads to n - 2 = 3 in
-        this case.
-        """
-
-        all_neighbors = list(local_search.ps1_gen(self.x))
-        assert len(all_neighbors) == 3
-
-    def test_ps2_returns_correct_num_neighbors(self):
-        """PS2 has n * (n - 1) / 2 swaps.
-        But since we fix the first element as origin, it leads to
-        (n - 1) * (n - 2) / 2 = 6 in this case.
-        """
-
-        all_neighbors = list(local_search.ps2_gen(self.x))
-        assert len(all_neighbors) == 6
-
-    def test_ps3_returns_correct_num_neighbors(self):
-        """PS3 has n * (n - 1) elements.
-        But since we fix the first element as origin, it leads to
-        (n - 1) * (n - 2) = 12 in this case.
-        """
-
-        all_neighbors = list(local_search.ps3_gen(self.x))
-        assert len(all_neighbors) == 12
-
-
 class TestLocalSearch:
     @pytest.mark.parametrize(
         "distance_matrix, expected_distance",
@@ -64,12 +33,12 @@ class TestLocalSearch:
         self, distance_matrix, expected_distance
     ):
         """
-        The _setup outputs the same input if provided, together with
+        The setup outputs the same input if provided, together with
         its objective value
         """
 
         x0 = [0, 2, 3, 1, 4]
-        x, fx = local_search._setup(distance_matrix, x0)
+        x, fx = local_search.setup(distance_matrix, x0)
 
         assert x == x0
         assert fx == expected_distance
@@ -80,11 +49,11 @@ class TestLocalSearch:
     )
     def test_setup_return_random_valid_solution(self, distance_matrix):
         """
-        The _setup outputs a random valid permutation if no
+        The setup outputs a random valid permutation if no
         initial solution is provided
         """
 
-        x, fx = local_search._setup(distance_matrix)
+        x, fx = local_search.setup(distance_matrix)
 
         assert set(x) == set(range(distance_matrix.shape[0]))
         assert fx
